Connect web and worker with RabbitMQ, add thumbnail processor
I need to use a queue in RabbitMQ to connect the web module and the worker module.
When an image is uploaded from the web project, a message containing the image properties
is sent to the queue. The worker listen to the queue, and when it receives
the message properties from the queue, it creates a thumbnail image with width and
height no more than 150px, the aspect ratio need to be the same as the original image.
The thumbnail is then stored in the same location, local dir or a separate bucket,
with a `_thumbnail` suffix in the file name.

NB. incorrectly add a dependency in the parent dependencyManagement
section. Manually removed.