Add local file processing for dev profiles in worker module
Just like the web project that supports falling back to local storage
when running in `dev` profile, the `FileProcessingService` should also
be aware of local storage. Extract the key methods in `FileProcessingService`
to a interface, and add another service implementation that processes the files
from the local file storage, when the dev profile is activated. Create abstract
parent class if necessary to ensure reusing common logics for different service
implementation. Ensure the message in the message queue has enough information
to identify if the file is from prod S3 or dev local storage.

Is `FileProcessingService` still needed, given in previous step it was refactored
to two different implementation: s3 for prod and local file for dev?