Properly handle the start and stop of Spring Boot app
Update the web and the worker Spring Boot app to use `ApplicationPidFileWriter`
so that they will write down the process PID to a file. Update the start.cmd script
to pass in the desired pid file path to the pid file writer in the app, and update
the stop.cmd script so that it will read the PID from that file and kill the process.

the `start /B cmd /c` command needs to be updated because currently there seems to
be some problem with the double quote escaping in the long command.

(Plus some manual updates)